extract genesis-config crate
