+++ sdk/Cargo.toml
@@ -40,6 +40,7 @@ full = [
     "dep:solana-ed25519-program",
     "dep:solana-compute-budget-interface",
     "dep:solana-keypair",
+    "dep:solana-genesis-config",
     "dep:solana-precompile-error",
     "dep:solana-precompiles",
     "dep:solana-presigner",
@@ -73,6 +74,7 @@ frozen-abi = [
     "solana-fee-structure/frozen-abi",
     "solana-account/frozen-abi",
     "solana-cluster-type/frozen-abi",
+    "solana-genesis-config/frozen-abi",
     "solana-inflation/frozen-abi",
     "solana-poh-config/frozen-abi",
     "solana-program/frozen-abi",
@@ -140,6 +142,7 @@ solana-frozen-abi = { workspace = true, optional = true, features = [
 solana-frozen-abi-macro = { workspace = true, optional = true, features = [
     "frozen-abi",
 ] }
+solana-genesis-config = { workspace = true, optional = true, features = ["bincode"] }
 solana-inflation = { workspace = true, features = ["serde"] }
 solana-instruction = { workspace = true }
 solana-keypair = { workspace = true, optional = true, features = [
